SegNet-Street-Segmentation/requirements.txt
tqdm
numpy==1.26.4
Pillow
opencv-python
tensorboard
pyyaml
matplotlib
